Rejected commit: Default

+++ src/main/java/com/leoman/entity/BaseEntity.java
@@ -22,7 +22,7 @@ public abstract class BaseEntity implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "create_date")
     private Long createDate;
@@ -30,14 +30,13 @@ public abstract class BaseEntity implements Serializable {
     @Column(name = "modify_date")
     private Long updateDate;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
-
     public Long getCreateDate() {
         return createDate;
     }
